Fix reason-phrase mangling
The legacy code reason-phrase requires const char*, which we no longer
have. We get an SBUf from the parser instead.
To avoid c_str() performance regressions or bloating the code with
StatusLine alterations we must take advantage of the default reason
phrases on output and RFC permission to ignore phrases delivered by the
server.