Fix force_request_body_continuation bug
When the force_request_body_continuation access list is configured squid is 
sends 100-Continue responses to all HTTP GET messages unless the admin
is very careful with the ACLs. This can be reproduced trivially with

  force_request_body_continuation allow all

We should not evaluate force_request_body_continuation if the request
does not include "Expect: 100-continue" header.

This is a Measurement Factory project