Pass event name from Cache() creator functions
We cannot use any kind of smart storage in the CredentialsCache class
because that would now result in the c-strign being deallocated when
there is potentially an event still existing and using it.

It cannot be dynamically created and deleted because ownership is split
between event and the cache itself. Neither is in a good position to
control a dynamic strings lifetime. So make it a goblal const char*.