Add support for GnuTLS and other non-OpenSSL logics to LockingPointer
This brings LockingPointer and TidyPointer back to the same API definition
for the reset() method. And ensures that cases where LockingPointer could
accidentally have been used with reset() now lock/unlock properly.

To do this a static global is defined to store the currently active Lock
for each allocated base T objects. For each type T the template Locks()
method needs to be instantiated in an appropriate place. If there is no
better place src/security/support.cc is used.