Support unified EUI format code in external_acl_type
Squid supports %>eui as a logformat specifier, which produces an EUI-48 
for IPv4 clients and an EUI-64 for IPv6 clients.  However, This is not 
allowed as a format specifier for the external ACLs, and you have to use 
%SRCEUI48 and %SRCEUI64 instead.  %SRCEUI48 is only useful for IPv4 
clients and %SRCEUI64 is only useful for IPv6 clients, so supporting 
both v4 and v6 is a bit messy.

Adds the %>eui specifier for external ACLs and behaves in the same way
as the logformat specifier.