Cleanup: update class fde and _fde_disk to C++11
This update moves the initialization of most class members into default
initizliation values. So clear() can be optimized to a fast emplace
initialization by the compiler and removing the need for memset().

NP: implicit initialization of function pointers and char* arrays is not
supported by C++11. So use of default constructor is still required for now.

Also,
- replace internal uses of fd_table macro with fde::Table.
- replace use of NULL with nullptr
- update some method declaratiosn to current coding style
- improve inline ability for noteUse() method
- const correctness for several inline methods
- documentation updates