Cleanup: remove unsused fd_debug_t and fdd_table
This tables used to store the file and line location where each FD was
closed from, but never read for any reports or logs.

Make the cache.log debugging for comm_close display that information
instead of storing it in memory. This way when debug of the comm layer
is enabled the closure information for each FD open/close cycle is
recorded and not just the place of last closure.

Also we save a few hundred KB of memory this table was using and all the
CPU cycles used to maintain it when debugging was not needed.