QA: allow test-suite to be run without a full build
The squid.conf processing tests have been assuming a full 'make check' was
run and generated a squid binary in the build directory.

This change allows callers to also run these tests on an arbitrary 'squid'
binary by using the command:
  make --eval="BIN_DIR=/path" -C test-suite squid-conf-tests

where /path is the path under which a squid binary already exists.