Replace tag with branch name
Reverting to the BZR way of doing things solely on branch name.
It turns out that all the git documentation saying to use tags
and indicating a tag named HEAD was best to use for latest code
snapshots was wrong.
So long as we consistently use the branch name in all git commands
they are applied to that branchs HEAD revision instead of the working
directory's checkout.