Auto-removed HERE from all debugs() calls. It is now unused.
This is a controversial change because it brings no runtime benefits,
creates significant cross-branch patch porting headaches and branch
maintenance overheads, but it also removes one source of persistent
bickering during code review and improves code consistency. Global
removal is the only practical way we can remove all HEREs -- restricting
removal of HEREs to changed-for-other-reasons lines would take forever.

The (simple) script to remove HEREs will be published separately because
it is not really needed in this source tree (which no longer uses HERE).