Abort abandoned and detached collapsed entries.
Before 9fff3bb, a collapsed entry was aborted during syncing, if
the shared entry became 'abandoned' (e.g., somebody marked it for
eventual deletion). This approach was flawed IMO, because it
aborted already attached store clients. The 9fff3bb fixed this, aborting
clients only if the shared entry became 'aborted' itself.

That fix brought another question: what if we have a collapsed entry,
not yet attached to a shared storage, and it gets 'abandon'
notification?
We will wait on this detached entry forever, because:
1. We are not able to attach to a shared storage now(the abandon marks
   for deletion that shared entry).
2. We do not abort our entry due to 9fff3bb change.

This change follows (2), aborting such abandoned and detached
collapsed entries.