DELETE should not affect already collapsed clients, pt. 2.
This is a clarification for 43121f6 fix.

When making an entry private, there is no need to expire it even if the
entry is unshareable. StoreEntry::expires is a rather weak protection
for entry reuse (e.g., it can be ignored due to max-stale directives).
Instead of expiration, the proper way is to rely on the entry-specific
attributes (such as KEY_PRIVATE, RELEASE_REQUEST and
shareableWhenPrivate), which should never be ignored.