Do not use 'Content-Range' header field for non-206/416 responses.
Fixed bugs 2821/3071.

Before this change, if Squid received 200 OK response with unexpected
Content-Range header field, it extracted and used that information,
applying it to offsets and length, thus randomly truncating message
body.

According to RFC 7233, only 206 and 416 responses prescribe the proper
meaning for 'Content-Range' header, there is no good reason for Squid to
use it for other responses. It is much better to relay it 'as-is',
letting client do required processing.