Do not assume exclusive ownership in MemStore::anchorEntry()
... but still be cautious about overwriting disk entry metadata.

The original code was probably designed to create a new StoreEntry from
scratch. Thus, it was important to set all the fields. Since then,
StoreEntry constructor got fixed, and, eventually, the anchorEntry()
method might be called for StoreEntry objects that already participate
in disk or other activities. The MemStore anchoring code should focus on
memory-related needs instead of assuming it is constructing a new
StoreEntry and duplicating constructor work.

This change was prompted by a stale/confusing comment left by branch
commit 35efd80.