Discovered a problem that makes strange isWriter() check necessary.
In theory, if we are completing a writing sequence, we have to be a
writer. The writer may not have a transients entry (if the StoreEntry
object went private), but if it does, then isWriter() has to be true.

In practice, we may be writing twice. When the second swapout starts,
the transient entry may already be in reading state. We never switch it
from reading to writing, even if we start writing again.

TODO: Consider removing Transients read/write state. Once created, the
Transients store entry probably does not need to be locked for writing,
so we do not need xitTable.io to track the store lock direction -- that
store lock can always be a read lock. The disk/memory swapout info in
the StoreEntry object can probably be used to distinguish readers from
local writers in Controller::syncCollapsed().