Removed check-this TODO.
Can empty anchors be locked? Yes, of course, because one must lock the
anchor before making it non-empty.

Should we mark empty locked (by others) anchors with waitingToBeFreed?
Yes, we should if we want to be conservative: A locked anchor may be
locked for writing. If the writing process writes the matching key, but
we do not mark the entry, then freeEntry() essentially fails to function
as documented. Without a lock, we cannot (reliably) tell whether there
is a match or even whether the anchor is empty.