Load Transients entry basics. Root.peek()ing RebuildState needs them.
Also do not store the KEY_PRIVATE flag in (Transient) anchors. To avoid
undoing on sharing failures, StoreEntry::setPublicKey() currently calls
Store::Root().addWriting() before forcePublicKey(), so the flag is still
set during StoreMapAnchor::set(). This is not so much a setPublicKey()
problem as it is an outdated private key handling -- we should just
replace private keys with (already used) nil keys.