When marking a Transient entry, call Controller::syncCollapsed().
Controller::syncCollapsed() was called if the entry was marked for
deletion in another worker, but same-worker readers could get stuck if
their worker initiated the eviction/Broadcast() because other workers,
seeing that the Transient entry is marked, did not Broadcast().

TODO: The "reentrant chains" concern may indicate that the Broadcast()
calling code is placed too low.

TODO: Deduplicate Transients::evictCached() and evictIfFound().