Commit 889fc47 was made to fix issue with Basic authentication and SSL bumping.
But after this commit we can no longer properly use http_access with
proxy_auth/proxy_auth_regex ACL because that type of ACL always return 1(match)
regardless of the conditions in the rules.

I add an authentication scheme check. If 889fc47 patch was made to fix the bug with
Basic authentication, then we should check whether this type of authentication is really used.