Repaired entries evicting after 5c09e58
Before this change (and after 5c09e58) a Transients entry sometimes
could not be properly evicted using Transients::evictCached() because
the StoreEntry had already been disconnected from the Transients cache,
while evicting shared memory and destroying memory object.
To overcome this, we need to evict the entry from Transients before
destroying memory object.
It may seem that doing so would violate the "transients invariant"
(every public StoreEntry must have the corresponding Transients entry),
for a very short period of time. However this is not a problem,
because the StoreEntry is not accessible via Controller interface
after it has been evicted form Transients.